expenses의 데이터 변화는 App.vue에서 되도록 설정 중
